+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### New features
 
+- Structs `Louis.Text.TrueConditionInterpolatedStringHandler` and `Louis.Text.FalseConditionInterpolatedStringHandler` are interpolated string handlers that only perform formatting if a given condition is true or false, respectively.  
+They can help implement, for example, conditional logging methods that take a condition and an interpolated string as a parameter, ensuring that string interpolation will only be performed if the result is actually used.
+
 ### Changes to existing features
 
 ### Bugs fixed in this release